Create separate class for debugger using prompt_toolkit
