+++ IPython/terminal/ptshell.py
@@ -26,6 +26,7 @@ from prompt_toolkit.styles import PygmentsStyle, DynamicStyle
 from pygments.styles import get_style_by_name, get_all_styles
 from pygments.token import Token
 
+from .debugger import TerminalPdb
 from .pt_inputhooks import get_inputhook_func
 from .interactiveshell import get_default_editor, TerminalMagics
 from .ptutils import IPythonPTCompleter, IPythonPTLexer
@@ -43,6 +44,7 @@ class TerminalInteractiveShell(InteractiveShell):
 
     pt_cli = None
     debugger_history = None
+    debugger_cls = TerminalPdb
 
     autoedit_syntax = Bool(False,
         help="auto editing of files with syntax errors.",
